Add Sebha screen and SettingsProvider
This commit adds a new Sebha screen to the app, which allows users to count their tasbih. It also adds a new SettingsProvider, which will be used to manage app settings in the future. The commit includes new image files for the Sebha screen, as well as localization updates for the new screen.